Refactor, UV transfer, and README updates
