+++ smplx_blender/data/smpl_v1_high.obj
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9df799fd5f28ffea722a9db6918ed70f75e259dbb19acb4f182df3dfa2548254
+size 17359075
