+++ smplx_blender/data/smpl_v1_low.obj
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:316a58e29b0aeeecd5abb10a39e38907260d4eaee9e207ceb06554c98841068f
+size 998884
