+++ smplx_blender/data/smpl_v1_medium.obj
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8ed9849156309dad0fb5e48b5f05fbd8c4d512d499ece69060321ffc01e49d7a
+size 4222471
