+++ smplx_blender/data/smplx_handposes.npz
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:dd70c2c33d12a3020522577dd8ab737c379a84c700463051ec5770ad990223d1
+size 808
