+++ smplx_blender/data/smplx_measurements_to_betas_female.json
@@ -1,0 +1,1 @@
+{"A": [[0.12746234828546227, 0.48113534457678786], [-0.07552187732092362, 3.1503577138098655], [0.010196594907119176, -0.8208804192164507], [0.004359222417471764, -0.3939571174023492], [0.0036761131870858745, -0.40365567816031855], [0.010458878120985993, -0.13955155886034984], [-0.0012401793877649607, -0.2862070245163085], [-0.0071686269888230555, 0.26920030166788667], [0.004207800329012893, 0.03054196582126984], [-0.0063954652405219135, 0.12072965190769482]], "B": [-23.020271977835808, -0.4257467833544609, 1.6513823917790753, 1.0241148039275119, 1.0885804302674715, -1.1986255890885535, 1.4398069468381105, 0.07663496090844153, -0.9249974381504901, 0.4519379539789322], "a": 0.0, "b": 1.0}
