+++ smplx_blender/data/smplx_measurements_to_betas_male.json
@@ -1,0 +1,1 @@
+{"A": [[-0.11789233533554878, -0.448592913436558], [-0.07320497685952516, 3.265025039523733], [-0.03048248339963719, 1.8967334887157248], [0.00805447152667242, -0.30600762798045145], [0.0035773932712879464, -0.4190802556602972], [0.004869741149622757, -0.22644903866227356], [0.008964351467087777, -0.8868353098937399], [-0.006049132890524903, 0.17533366043620105], [-0.005547104891185135, 0.49849734830065184], [-0.0014086631203506686, 0.29232931393919054]], "B": [23.049146483674704, -1.179793795471003, -2.765917837117154, -0.2619036434678064, 1.213111118104932, 0.16666576523578058, 2.243161996211576, 0.3429807759493095, -1.0618240192839958, -0.9416353818479539], "a": 0.0, "b": 1.0}
