+++ smplx_blender/data/smplx_texture_f_alb.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2ea01f4fd7c64af99de3f2446dfe446758f8083c9957fcda22a37b7204812c59
+size 18160879
