+++ smplx_blender/data/smplx_texture_m_alb.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:5adf6ef6bd22c3f325fec8281d4cfdbc904abf029409db5ec7cc9a4cec26bb58
+size 20285692
