+++ smplx_blender/data/smplx_texture_rainbow.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e2d0e3392a0697483561336c975bbd6065b6b5335c1b5533ae3c83bf277a51b9
+size 3636050
