+++ smplx_blender/data/uv_high.pkl
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a4bfd50032c632d8c84cdfb7872dd67e355616f0f5656f07e7d80405c6b2b042
+size 2484615
