+++ smplx_blender/data/uv_low.pkl
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2bedf053055cf7bab6aa7005d40e550d9c294208eed482a8c71b188b4b6c9c63
+size 166038
