+++ smplx_blender/data/uv_medium.pkl
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b8ede1bb8ae7eb2818c09e871d1944f223cc5f5270e6a2c0821234346410bf75
+size 635579
